requirements.txt
Flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.3.2
mysql-connector-python==8.2.0
gunicorn==21.2.0